ShapeUtils: Move triangulation code to separate file
